slack/requirements.txt
flask==2.3.3
pyyaml==6.0.1
pydantic==1.10.13
structlog==23.2.0